# New log circular buffer