taken from https://go-review.googlesource.com/c/go/+/478200/3/src/log/slog/internal/benchmarks/handlers.go#121