@komuw 2023-03-27
The alternative is slower:

```
benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/komuw/ong/log
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
            │   old.txt   │               new.txt                │
            │   sec/op    │    sec/op     vs base                │
CircleBuf-8   12.23n ± 0%   18.06n ± 13%  +47.77% (p=0.000 n=20)

            │  old.txt   │            new.txt             │
            │    B/op    │    B/op     vs base            │
CircleBuf-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal

            │  old.txt   │            new.txt             │
            │ allocs/op  │ allocs/op   vs base            │
CircleBuf-8   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal
```